Created structure for Github Client APIs
