+++ app/services/code_owners/request.rb
@@ -2,7 +2,7 @@ module CodeOwners
   class Request < BaseService
     def call
       Project.find_each do |project|
-        content_file = GithubRepositoryClient.new(project).code_owners
+        content_file = GithubClient::Repository.new(project).code_owners
         next if content_file.empty?
 
         code_owners = CodeOwners::FileHandler.call(content_file)
