+++ app/services/processors/open_source_project_views_updater.rb
@@ -34,7 +34,7 @@ module Processors
     end
 
     def views_payload
-      GithubRepositoryClient.new(project).repository_views
+      GithubClient::Repository.new(project).views
     end
 
     def update_views_metric(timestamp, views)
