+++ spec/services/github_client/repository_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe GithubRepositoryClient do
+RSpec.describe GithubClient::Repository do
   describe '#code_owners' do
     let(:project) { create(:project, name: 'rs-code-metrics') }
 
@@ -21,7 +21,7 @@ RSpec.describe GithubRepositoryClient do
     end
   end
 
-  describe '#repository_views' do
+  describe '#views' do
     let(:project) { create(:project) }
     let(:repository_views_payload) { create(:repository_views_payload) }
 
@@ -29,7 +29,7 @@ RSpec.describe GithubRepositoryClient do
       before { stub_successful_repository_views(project, repository_views_payload) }
 
       it 'returns the views hash of that project on Github' do
-        expect(described_class.new(project).repository_views).to eq repository_views_payload
+        expect(described_class.new(project).views).to eq repository_views_payload
       end
     end
 
@@ -37,7 +37,7 @@ RSpec.describe GithubRepositoryClient do
       before { stub_failed_repository_views(project) }
 
       it 'raises an exception' do
-        expect { described_class.new(project).repository_views }.to raise_error Faraday::ClientError
+        expect { described_class.new(project).views }.to raise_error Faraday::ClientError
       end
     end
   end
